extract to file
