+++ lib/deno_postgres_interop.dart
@@ -2,6 +2,7 @@
 library;
 
 export 'src/client.dart';
+export 'src/isolation_level.dart';
 export 'src/query_client.dart';
 export 'src/query_object_result.dart';
 export 'src/transaction.dart';
