+++ lib/src/isolation_level.dart
@@ -1,0 +1,8 @@
+enum IsolationLevel {
+  readCommited,
+  repeatableRead,
+  serializable;
+
+  static IsolationLevel parse(String string) =>
+      values.firstWhere((e) => e.name == string);
+}
